move endpoints.js into data folder
