+++ src/components/EndpointExplorer.js
@@ -4,7 +4,7 @@ import {connect} from 'react-redux';
 import {EndpointPicker} from './EndpointPicker';
 import {EndpointSetup} from './EndpointSetup';
 import {EndpointResult} from './EndpointResult';
-import {getEndpoint} from '../endpoints';
+import {getEndpoint} from '../data/endpoints';
 import UriTemplates from 'uri-templates';
 
 class EndpointExplorer extends React.Component {
